+++ packages/lumx-react/src/components/list/ListItem.stories.tsx
@@ -4,6 +4,7 @@ import { withWrapper } from '@lumx/react/stories/decorators/withWrapper';
 import { CustomLink } from '@lumx/react/stories/utils/CustomLink';
 import { withCombinations } from '@lumx/react/stories/decorators/withCombinations';
 import { getSelectArgType } from '@lumx/react/stories/controls/selectArgType';
+import { isEqual } from '@lumx/react/utils/object/isEqual';
 import { ListItem } from './ListItem';
 
 const sizes: ListItemSize[] = [Size.tiny, Size.regular, Size.big];
@@ -21,51 +22,8 @@ export default {
 /**
  * Default list item with text
  */
-export const NonClickable = {
+export const Default = {
     args: { children: 'List item' },
-};
-
-/**
- * Button list item (onClick)
- */
-export const Button = {
-    args: {
-        children: 'List item button',
-    },
-    argTypes: {
-        onItemSelected: { action: true },
-    },
-};
-
-/**
- * Disabled button
- */
-export const ButtonDisabled = {
-    ...Button,
-    args: {
-        ...Button.args,
-        isDisabled: true,
-    },
-};
-
-/**
- * Link list item (href)
- */
-export const Link = {
-    args: {
-        linkProps: { href: '#' },
-        children: 'List item link',
-    },
-};
-
-/**
- * Disabled link
- */
-export const LinkDisabled = {
-    args: {
-        ...Link.args,
-        isDisabled: true,
-    },
 };
 
 /**
@@ -79,22 +37,30 @@ export const CustomLink_ = {
 };
 
 /**
- * Combination of size and states
+ * Combination of all states (size, clickable and non-clickable states)
  */
-export const SizeAndStates = {
-    ...Button,
+export const AllStates = {
+    args: { children: 'List item' },
     decorators: [
-        withWrapper({}, List),
         withCombinations({
             combinations: {
                 rows: { key: 'size', options: sizes },
                 cols: {
                     Default: {},
                     Disabled: { isDisabled: true },
+                    'ARIA Disabled': { 'aria-disabled': true },
                     Selected: { isSelected: true },
                     Highlighted: { isHighlighted: true },
                 },
+                sections: {
+                    Default: {},
+                    'As button': { onItemSelected: () => {} },
+                    'As link': { linkProps: { href: '#' } },
+                },
             },
+            // Only keep size variants for non clickable list items
+            excludeCombination: (props) =>
+                !props.onItemSelected && !props.linkProps?.href && !isEqual(Object.keys(props), ['size', 'children']),
         }),
     ],
 };