+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Checkbox`: handle `aria-disabled=true` for an accessible disabled state
 -   `Chip`: handle `aria-disabled=true` for an accessible disabled state
 -   `Link`: handle `aria-disabled=true` for an accessible disabled state
+-   `ListItem`: handle `aria-disabled=true` for an accessible disabled state
 
 ### Fixed